move Clock class to utils and move test setup to use ISO-8601 formatting
